requirements.txt
Pillow
pandas
python-dateutil
scrapy
